+++ src/wavefunctions/hf/frozen_orbital_hf.F90
@@ -70,6 +70,11 @@ contains
 !
       if (wf%frozen_hf_mos) call wf%remove_frozen_hf_orbitals()
 !
+      if (input%is_keyword_present('print orbitals', 'system') .and. &
+         (wf%frozen_core .or. wf%frozen_hf_mos)) then
+         call wf%write_orbitals("nonfrozen")
+      end if
+!
    end subroutine prepare_mos_hf
 !
 !